raspberry_pi_yolo/requirements.txt
# Core ML Framework
ultralytics>=8.0.0,<9.0.0
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0

# Computer Vision
opencv-python>=4.8.0,<5.0.0
Pillow>=10.0.0,<11.0.0

# Data Processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
pyyaml>=6.0,<7.0

# System Monitoring (for Pi optimization)
psutil>=5.9.0,<6.0.0

# Progress and Logging
tqdm>=4.65.0,<5.0.0
loguru>=0.7.0,<1.0.0

# Optional: ONNX Runtime (for model optimization)
# Uncomment if wanted ONNX optimization on Pi
# onnx>=1.14.0,<2.0.0
# onnxruntime>=1.15.0,<2.0.0

# Optional: GPU Monitoring
# Uncomment if wanted GPU monitoring on PC
# GPUtil>=1.4.0,<2.0.0 